Merged with version 3.0

+++ lib/JFFNetwork/DownloadManager/Detail/NSMutableDictionary+DownloadingFileInfo.m
@@ -27,7 +27,8 @@
         
         return [fileLength unsignedLongLongValue];
     }
-    return NSURLResponseUnknownLength;
+    
+    return (unsigned long long)NSURLResponseUnknownLength;
 }
 
 + (void)setFileLength:(unsigned long long)fileLength
